Adapt existing tests to new properties
Moved sort, search, and filter tests into CaseDetailScreenSortAndFilter

Jira ticket: CAMS-255
