+++ user-interface/src/case-detail/CaseDetailScreen.test.tsx
@@ -1,16 +1,10 @@
 import { BrowserRouter, MemoryRouter } from 'react-router-dom';
 import { describe } from 'vitest';
 import { render, waitFor, screen, queryByTestId, fireEvent } from '@testing-library/react';
-import {
-  CaseDetail,
-  docketSorterClosure,
-  applySortAndFilters,
-  getSummaryFacetList,
-} from './CaseDetailScreen';
+import { CaseDetail, docketSorterClosure } from './CaseDetailScreen';
 import { getCaseNumber } from '@/lib/utils/formatCaseNumber';
 import {
   CaseDetailType,
-  CaseDocket,
   CaseDocketEntry,
   Debtor,
   DebtorAttorney,
@@ -772,167 +766,6 @@ describe('Case Detail screen tests', () => {
       const fn = docketSorterClosure('Oldest');
       const expectedValue = -1;
       expect(fn(left, right)).toEqual(expectedValue);
-    });
-  });
-
-  describe('sort, search, and filter tests', () => {
-    const testCaseDocketEntries: CaseDocket = [
-      {
-        sequenceNumber: 2,
-        documentNumber: 1,
-        dateFiled: '2023-05-07T00:00:00.0000000',
-        summaryText: 'Add Judge',
-        fullText: 'Docket entry number 1.',
-      },
-      {
-        sequenceNumber: 3,
-        dateFiled: '2023-05-07T00:00:00.0000000',
-        summaryText: 'Motion',
-        fullText: 'Docket entry number 2.',
-      },
-      {
-        sequenceNumber: 4,
-        documentNumber: 2,
-        dateFiled: '2023-07-07T00:00:00.0000000',
-        summaryText: 'Add Attorney',
-        fullText: 'Docket entry number 3.',
-        documents: [
-          {
-            fileLabel: '0-0',
-            fileSize: 1000,
-            fileExt: 'pdf',
-            fileUri: 'https://somehost.gov/pdf/0000-111111-3-0-0.pdf',
-          },
-        ],
-      },
-      {
-        sequenceNumber: 5,
-        dateFiled: '2023-05-07T00:00:00.0000000',
-        summaryText: 'Motion',
-        fullText: 'Docket entry number 4.',
-      },
-    ];
-
-    test('should filter the list of docket entries per the search text', async () => {
-      const filteredDocketEntries = applySortAndFilters(testCaseDocketEntries, {
-        searchString: 'number 2',
-        selectedFacets: [],
-        sortDirection: 'Oldest',
-      });
-
-      expect(filteredDocketEntries?.length).toEqual(1);
-      const actualEntry = filteredDocketEntries ? filteredDocketEntries[0] : null;
-      expect(actualEntry).toEqual(testCaseDocketEntries[1]);
-    });
-
-    test('should filter the list of docket entries per the selected facets', async () => {
-      const filteredDocketEntries = applySortAndFilters(testCaseDocketEntries, {
-        searchString: '',
-        selectedFacets: [
-          testCaseDocketEntries[1].summaryText,
-          testCaseDocketEntries[3].summaryText,
-        ],
-        sortDirection: 'Oldest',
-      });
-
-      expect(filteredDocketEntries?.length).toEqual(2);
-      const actualEntriesOne = filteredDocketEntries ? filteredDocketEntries[0] : null;
-      expect(actualEntriesOne).toEqual(testCaseDocketEntries[1]);
-      const actualEntriesTwo = filteredDocketEntries ? filteredDocketEntries[1] : null;
-      expect(actualEntriesTwo).toEqual(testCaseDocketEntries[3]);
-    });
-
-    test('should sort the list of docket entries oldest first', async () => {
-      const youngestEntry = testCaseDocketEntries[2];
-      const middleEntry = testCaseDocketEntries[1];
-      const oldestEntry = testCaseDocketEntries[0];
-
-      const docketEntries = testCaseDocketEntries.slice(0, 3);
-      const filteredDocketEntries = applySortAndFilters(docketEntries, {
-        searchString: '',
-        selectedFacets: [],
-        sortDirection: 'Oldest',
-      });
-
-      expect(filteredDocketEntries?.length).toEqual(3);
-      const first = filteredDocketEntries ? filteredDocketEntries[0] : null;
-      const second = filteredDocketEntries ? filteredDocketEntries[1] : null;
-      const third = filteredDocketEntries ? filteredDocketEntries[2] : null;
-      expect(first).toEqual(oldestEntry);
-      expect(second).toEqual(middleEntry);
-      expect(third).toEqual(youngestEntry);
-    });
-
-    test('should sort the list of docket entries newest first', async () => {
-      const youngestEntry = testCaseDocketEntries[2];
-      const middleEntry = testCaseDocketEntries[1];
-      const oldestEntry = testCaseDocketEntries[0];
-
-      const docketEntries = testCaseDocketEntries.slice(0, 3);
-      const filteredDocketEntries = applySortAndFilters(docketEntries, {
-        searchString: '',
-        selectedFacets: [],
-        sortDirection: 'Newest',
-      });
-
-      expect(filteredDocketEntries?.length).toEqual(3);
-      const first = filteredDocketEntries ? filteredDocketEntries[0] : null;
-      const second = filteredDocketEntries ? filteredDocketEntries[1] : null;
-      const third = filteredDocketEntries ? filteredDocketEntries[2] : null;
-      expect(first).toEqual(youngestEntry);
-      expect(second).toEqual(middleEntry);
-      expect(third).toEqual(oldestEntry);
-    });
-
-    test('should sort facets in call to getDocumentSummaryFacets', async () => {
-      const testFacets = new Map([
-        [
-          'Motion for Joint Administration',
-          {
-            text: 'Motion for Joint Administration',
-            count: 5,
-          },
-        ],
-        [
-          'Add Judge',
-          {
-            text: 'Add Judge',
-            count: 2,
-          },
-        ],
-        [
-          'Case Association - Joint Administration',
-          {
-            text: 'Case Association - Joint Administration',
-            count: 2,
-          },
-        ],
-        [
-          'Order Re: Motion for Joint Administration',
-          {
-            text: 'Order Re: Motion for Joint Administration',
-            count: 1,
-          },
-        ],
-      ]);
-      const expectedFacets = [
-        { value: 'Add Judge', label: 'Add Judge (2)' },
-        {
-          value: 'Case Association - Joint Administration',
-          label: 'Case Association - Joint Administration (2)',
-        },
-        {
-          value: 'Motion for Joint Administration',
-          label: 'Motion for Joint Administration (5)',
-        },
-        {
-          value: 'Order Re: Motion for Joint Administration',
-          label: 'Order Re: Motion for Joint Administration (1)',
-        },
-      ];
-
-      const resultFacets = getSummaryFacetList(testFacets);
-      expect(resultFacets).toStrictEqual(expectedFacets);
     });
   });
 });
